A working model with direction outside the stan model
Next steps: Pare down the multiple models to two: width parameterized, and w parameterized, with everything else the same and fully featured. Pass des in as data, rename in increase/decrease. Then, new parsing of prior config file to generate the values for the prior distribution and pass them in as data.